+++ packages/typespec-rust/src/codegen/enums.ts
@@ -63,7 +63,6 @@ export function emitEnums(crate: rust.Crate, context: Context): helpers.Module |
   // emit impls as required
   for (const rustEnum of crate.enums) {
     body += context.getTryFromForRequestContent(rustEnum, use);
-    body += context.getModelImplForType(rustEnum, use);
   }
 
   let content = helpers.contentPreamble();