+++ packages/typespec-rust/test/spector/azure/resource-manager/common-properties/tests/common_properties_managed_identity_client_test.rs
@@ -2,7 +2,7 @@
 //
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-use azure_core::credentials::{AccessToken, TokenCredential};
+use azure_core::credentials::{AccessToken, TokenCredential, TokenRequestOptions};
 use azure_core::date::OffsetDateTime;
 use azure_core::Result;
 use spector_armcommon::models::{
@@ -26,7 +26,11 @@ impl FakeTokenCredential {
 
 #[async_trait::async_trait]
 impl TokenCredential for FakeTokenCredential {
-    async fn get_token(&self, _scopes: &[&str]) -> Result<AccessToken> {
+    async fn get_token(
+        &self,
+        _scopes: &[&str],
+        _options: Option<TokenRequestOptions>,
+    ) -> Result<AccessToken> {
         Ok(AccessToken::new(
             self.token.clone(),
             OffsetDateTime::now_utc(),